Crate snowbridge_milagro_bls

Modules

BLSCurve

Structs

AggregatePublicKey

Allows for the adding/combining of multiple BLS PublicKeys.

AggregateSignature

Allows for the adding/combining of multiple BLS Signatures.

Keypair

A helper which stores a BLS public and private key pair.

PublicKey

A BLS public key.

SecretKey

A BLS secret key.

Signature

Enums

AmclError

Constants

G1_BYTES

The required number of bytes for a compressed G1 point

G2_BYTES

The required number of bytes for a compressed G2 point

SECRET_KEY_BYTES

The required number of bytes for a secret key