Module snowbridge_milagro_bls::BLSCurve::bls381::proof_of_possession

Constants

DST_G1

Domain Separation Tag for signatures on G1

DST_G2

Domain Separation Tag for signatures on G2

DST_POP_G1

Domain Separation Tag for PopProve and PopVerify for signatures on G1

DST_POP_G2

Domain Separation Tag for PopProve and PopVerify with signatures on G2

G1_BYTES

The required number of bytes for a compressed G1 point

G2_BYTES

The required number of bytes for a compressed G2 point

SECRET_KEY_BYTES

The required number of bytes for a secret key

Functions

aggregate_g1

Aggregate

aggregate_g2

Aggregate

aggregate_verify_g1

Proof of Possession - AggregateVerify

aggregate_verify_g2

Proof of Possession - AggregateVerify

fast_aggregate_verify_g1

Proof of Possession - FastAggregateVerify

fast_aggregate_verify_g2

Proof of Possession - FastAggregateVerify

key_generate

Proof of Possession - KeyGenerate

key_pair_generate_g1

Generate key pair - (secret key, public key)

key_pair_generate_g2

Generate key pair - (secret key, public key)

pop_prove_g1

Proof of Possession - PopProve

pop_prove_g2

Proof of Possession - PopProve

pop_verify_g1

Proof of Possession - PopVerify

pop_verify_g2

Proof of Possession - PopVerify

secret_key_to_public_key_g1

Secret Key To Public Key

secret_key_to_public_key_g2

Secret Key To Public Key

sign_g1

Proof of Possession - Sign

sign_g2

Proof of Possession - Sign

verify_g1

Proof of Possession - Verify

verify_g2

Proof of Possession - Verify