Module snowbridge_milagro_bls::BLSCurve::mpin

Constants

BAD_PARAMS

BAD_PIN

EFS

EGS

INVALID_POINT

MAXPIN

PBLEN

SHA256

SHA384

SHA512

TRAP

TS

WRONG_ORDER

Functions

client_1

Implement step 1 on client side of MPin protocol

client_2

Implement step 2 on client side of MPin protocol

client_key

Calculate common key on client side wCID = w.(A+AT)

decoding

encoding

extract_factor

Extract factor from TOKEN for identity CID

extract_pin

Extract PIN from TOKEN for identity CID

get_client_permit

Time Permit CTT=S*(date|H(CID)) where S is master secret

get_client_secret

Client secret CST=S*H(CID) where CID is client ID and S is master secret CID is hashed externally

get_g1_multiple

W=xH(G); if RNG == NULL then X is passed in if RNG != NULL the X is passed out if type=0 W=xG where G is point on the curve, else W=x*M(G), where M(G) is mapping of octet G to point on the curve

get_server_secret

Extract Server Secret SST=S*Q where Q is fixed generator in G2 and S is master secret

get_y

Generate Y = H(epoch, xCID/xID)

hash_all

Hash the M-Pin transcript - new

hash_id

kangaroo

Pollards kangaroos used to return PIN error

precompute

Functions to support M-Pin Full

random_generate

create random secret S

recombine_g1

R=R1+R2 in group G1

recombine_g2

W=W1+W2 in group G2

restore_factor

Restore factor to TOKEN for identity CID

server_1

Outputs H(CID) and H(T|H(CID)) for time permits. If no time permits set HID=HTID

server_2

Implement step 2 of MPin protocol on server side

server_key

calculate common key on server side Z=r.A - no time permits involved